refactor to separate patrol reward module
