+++ nekoyume/Assets/_Scripts/UI/Widget/Popup/EventRewardPopup.cs
@@ -1,0 +1,7 @@
+namespace Nekoyume.UI
+{
+    public class EventRewardPopup : PopupWidget
+    {
+
+    }
+}
