+++ nekoyume/Assets/_Scripts/UI/Widget/Popup/EventRewardPopup.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3ef4c20ac0124810ac14d1e933f83ae4
+timeCreated: 1729581583
